REF: implement array_algos.take
